+++ tests/runtime/azure/test_azure_runtime.py
@@ -42,7 +42,7 @@ from qbraid.runtime.azure import AzureQuantumDevice, AzureQuantumJob
 from qbraid.runtime.azure.io_format import InputDataFormat, OutputDataFormat
 from qbraid.runtime.azure.provider import AzureQuantumProvider, serialize_pulser_input
 from qbraid.runtime.azure.result_builder import AzureResultBuilder
-from qbraid.runtime.postprocess import normalize_counts
+from qbraid.runtime.postprocess import normalize_data
 
 pytestmark = pytest.mark.filterwarnings("ignore:Unrecognized input data format:UserWarning")
 
@@ -847,7 +847,7 @@ def test_azure_quantum_result_counts(
         return_value=mock_builder_ionq_results["results"],
     ):
         raw_counts = azure_result_builder.get_counts()
-        formatted_counts = normalize_counts(raw_counts)
+        formatted_counts = normalize_data(raw_counts)
     assert raw_counts == formatted_counts == {"000": 50, "111": 50}